secured all enpoints and added authentication with OAuth and JWT

+++ routes/events.py
@@ -9,6 +9,8 @@ from beanie import PydanticObjectId
 from database.connection import Database
 from models.events import Event
 from typing import List
+from auth.authenticate import authenticate
+
 event_database = Database(Event)
 event_router = APIRouter(
  tags=["Events"]
@@ -32,7 +34,8 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
  return event
 
 @event_router.post("/new")
-async def create_event(body: Event) -> dict:
+async def create_event(body: Event,user:str=Depends(authenticate)) -> dict:
+    body.creator = user
     event_database.save(body)
     return {
     "message": "Event created successfully"
@@ -49,7 +52,14 @@ async def create_event(body: Event) -> dict:
 
 
 @event_router.put("/{id}", response_model=Event)
-async def update_event(id: PydanticObjectId, body: EventUpdate):
+async def update_event(id: PydanticObjectId, body: EventUpdate,user: str= Depends(authenticate)):
+    event = await event_database.get(id)
+    if event.creator != user:
+                raise HTTPException(
+        status_code=status.HTTP_400_BAD_REQUEST,
+        detail="Operation not allowed"
+        )
+
     updated_event = await event_database.update(id, body)
     if not updated_event:
         raise HTTPException(
@@ -59,7 +69,13 @@ async def update_event(id: PydanticObjectId, body: EventUpdate):
     return updated_event
 
 @event_router.delete("/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
+async def delete_event(id: PydanticObjectId,user:str=Depends(authenticate)) -> dict:
+ event = event_database.get(id)
+ if event.creator != user:
+    raise HTTPException(
+        status_code = status.HTTP_404_NOT_FOUND,
+        detail = "Cant delete event(permission denied)"
+    )
  event = await event_database.delete(id)
  if not event:
         raise HTTPException(
